backend/requirements.txt
fastapi[standard]>=0.110.0
beanie>=1.26.0
motor<3.4.0
pymongo<4.7
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
python-dotenv>=1.0.1
langchain>=0.2.0
langchain-openai>=0.1.0
uuid>=1.30
pyjwt>=2.8.0
crawl4ai>=0.4.0
playwright>=1.41.0
pypdf>=4.0.0
faiss-cpu>=1.7.0
langchain-community>=0.2.0
